Support current_task for batch tasks
